requirements.txt
asgiref==3.2.3
dj-database-url==0.5.0
dj-static==0.0.6
Django==3.0.2
django-froala-editor==3.1.0
django-heroku==0.3.1
django-tinymce4-lite==1.7.5
gunicorn==20.0.4
Pillow==7.0.0
psycopg2==2.8.4
dj-database-url==0.5.0
psycopg2-binary==2.8.4
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
static3==0.7.0
urllib3==1.25.8
whitenoise==5.0.1
django-widget-tweaks==1.4.8
